refactorize language
