--- src/lib/language/index.ts
@@ -1,7 +1,0 @@
-export * from '../model/operands'
-export * from '../manager/operandManager'
-export * from './language'
-export * from '../manager/dmlBuilder'
-export * from '../manager/ddlBuilder'
-export * from './languages'
-export * from './dialect'
