+++ src/lib/manager/index.ts
@@ -10,3 +10,8 @@ export * from './dmlBuilder'
 export * from './routing'
 export * from './helper'
 export * from './operandManager'
+export * from './language'
+export * from './dmlBuilder'
+export * from './ddlBuilder'
+export * from './languages'
+export * from './dialect'
