+++ src/lib/model/index.ts
@@ -9,3 +9,4 @@ export * from './sentenceInfo'
 export * from './contextInfo'
 export * from './error'
 export * from './metadata'
+export * from './operands'
